Split out the Objective-C parts of the newtype conformance test
